feat: add ProductOrder component
